ctxseg/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
chex==0.1.90
click==8.3.0
ctxseg @ git+https://github.com/jiyuuchc/ctxseg@6fe2139a3c317f30c32996748127abd7ce8ed308
einops==0.8.1
etils==1.13.0
flax==0.10.7
fsspec==2025.9.0
grpcio==1.75.1
humanize==4.13.0
imageio==2.37.0
importlib_resources==6.5.2
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
lazy_loader==0.4
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.13.1.26
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.4.5
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.25
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
PyYAML==6.0.3
rich==14.1.0
scikit-image==0.25.2
scipy==1.15.3
shellingham==1.5.4
simplejson==3.20.2
tensorstore==0.1.77
tifffile==2025.5.10
toolz==1.0.0
treescope==0.1.10
typer==0.19.2
typing_extensions==4.15.0
zipp==3.23.0
